change switchedShunt to shuntCompensator

+++ ieee.odm_pss/src/org/ieee/pes/odm/pss/model/ODMModelParser.java
@@ -224,7 +224,7 @@ public class ODMModelParser {
 	 * @param id
 	 * @return
 	 */
-	public BusRecordXmlType addNewBaseCaseBus(String id, int number) throws Exception {
+	public BusRecordXmlType addNewBaseCaseBus(String id, long number) throws Exception {
 		BusRecordXmlType bus = getStudyCase().getBaseCase().getBusList().addNewBus();
 		bus.setId(id);
 		bus.setNumber(number);
